+++ charts/src/main/java/org/patternfly/chart/utilization/DonutUtilizationElement.java
@@ -21,12 +21,16 @@ import org.patternfly.chart.Threshold;
 
 import elemental2.core.JsArray;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 
 @JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "HTMLElement")
 public class DonutUtilizationElement extends ChartElement {
 
+    @JsProperty
     public Data data;
+    @JsProperty
     public boolean invert;
+    @JsProperty
     public JsArray<Threshold> thresholds;
 }