+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Add missing externs for the custom chart element.
+
 ## [0.4.0] - 2025-12-13
 
 ### Added